requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
pytz
pydantic[email]
pytest
httpx